requirements.txt
streamlit
tensorflow====2.12.0
pillow
numpy
pandas
altair
